% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC_HD.R
\name{PTC.NPMLE.ALasso.fit}
\alias{PTC.NPMLE.ALasso.fit}
\title{Variable selection in promotion time cure model (via adaptive Lasso)
based on the classical nonparametric maximum likelihood estimator and adaptive Lasso penalty}
\usage{
PTC.NPMLE.ALasso.fit(
  yobs,
  delta,
  X,
  LinkFunc = list(D0 = exp, D1 = exp),
  tunpara.num = 50,
  tunpara.min.ratio = 1e-04,
  tunpara = NULL,
  evatype = c("AIC", "BIC")[2],
  learnrate = 0.1,
  maxit = 1000,
  eps = 1e-06,
  threshold = 1e-05,
  trace = TRUE
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{tunpara.num}{the number of tuning parameter values that will be fitted at.}

\item{tunpara.min.ratio}{Smallest value for tuning parameter, as a fraction of tunpara.max, the (data derived) entry value (i.e. the smallest value for which all coefficients are zero).}

\item{tunpara}{A user supplied \code{tunpara} sequence. Typical usage is to have the program compute its own \code{tunpara} sequence based on \code{tunpara.num} and \code{tunpara.min.ratio}. Supplying a value of \code{tunpara} overrides this.}

\item{evatype}{the information criterion that will be used in the selection of optimal tuning parameter.}

\item{learnrate}{the learning rate in solving the optimization problem.}

\item{maxit}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default \code{maxit = 5e3}.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{threshold}{a threshold value. The deafult value is \code{1e-5}. If the absolute value of the each component of coefficients is smaller than \code{threshold}, it will be converted into zero.}

\item{trace}{a logical value. The default specification is \code{TRUE}, indicating that several useful information about the current process of fitting will be displayed.}
}
\value{
The fitted results are returned (a list). The estimates of regression coefficients is \code{res}.
}
\description{
Conduct variable selection using promotion time cure model based on the nonparametric maximum likelihood estimation (NPMLE) procedure and adaptive Lasso penalty.
In other words, this function can analyze survival data with a cure fraction and selection the underlying useful variables.
}
\examples{
### ==== An example for doing online variable selection via the promotion time cure model ==== ###

## ---- generate the simulated dataset ---- ##
sdata.full  <- sdata.PTC.High(N=500)
yobs  <- sdata.full$sdata$yobs
delta <- sdata.full$sdata$delta
X     <- as.matrix(sdata.full$sdata$X)
rm(sdata.full)

## ---- fit the promotion time cure model ---- ##
# estimator of the regression coefficients
fit.PTC <- PTC.NPMLE.ALasso.fit(yobs,delta,X)
print(fit.PTC$res)
# estimator of the baseline distribution at certain points
tm <- c(0.1,0.3,0.5,0.7,0.9)
Est.Lam0t <- PTC.NPMLE.Lam(tm=tm,bet=fit.PTC$res[,1],yobs=yobs,delta=delta,X=X)
print(data.frame(Est=Est.Lam0t,row.names=tm))

}
