% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC.R
\name{PTC.NPMLE.Beta.LogLik}
\alias{PTC.NPMLE.Beta.LogLik}
\title{Loss function for bet in NPMLE}
\usage{
PTC.NPMLE.Beta.LogLik(
  bet,
  yobs,
  delta,
  XI,
  LinkFunc = list(D0 = exp, D1 = exp),
  w = NULL
)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{XI}{a matrix of covariates with 1 as the first column, that is, \code{XI=cbind(1,X)}.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{w}{weights for the weighted bootstrap method. The default is \code{NULL}, and it will be treatead as equal weights (\code{w=rep(1,length(yobs))}).}
}
\description{
Loss function for the promotion cure model (Profiled log-likelihood function in NPMLE).
}
