% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC.R
\name{PTC.NPMLE.fit}
\alias{PTC.NPMLE.fit}
\title{Promotion time cure model based on the nonparametric maximum likelihood estimation procedure}
\usage{
PTC.NPMLE.fit(
  yobs,
  delta,
  X,
  LinkFunc = list(D0 = exp, D1 = exp),
  boots = FALSE,
  nboot = 100,
  eps = 1e-06,
  maxit = 5000,
  simplify = TRUE
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{boots}{a logical value. The default specification is \code{TRUE}, indicating that the standard errors will be provided using bootstrap.}

\item{nboot}{specifies the number of bootstrap sampling. The default \code{nboot = 100}.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{maxit}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default \code{maxit = 5e3}.}

\item{simplify}{a logical value. The default specification is \code{TRUE}, indicating that whether we will use the internal \code{optim} function in \code{R} to solve the method.}

\item{trace}{a binary variable that determines whether to display the information for each bootstrap iteration.}
}
\value{
The fitted results are returned (a list). The estimates of regression coefficients is \code{res}.
}
\description{
Fit promotion time cure model based on the nonparametric maximum likelihood estimation (NPMLE) procedure.
More details about the classical NPMLE method can be found in Portier et al. (2017).
}
\examples{
### ==== An example for fitting the promotion time cure model ==== ###

## ---- generate the simulated dataset ---- ##
sdata.full  <- sdata.PTC(N=200)
yobs  <- sdata.full$sdata$yobs
delta <- sdata.full$sdata$delta
X     <- as.matrix(sdata.full$sdata$X)
rm(sdata.full)

## ---- fit the promotion time cure model ---- ##
# estimator of the regression coefficients
fit.PTC <- PTC.NPMLE.fit(yobs,delta,X)
print(fit.PTC$res)
# estimator of the baseline distribution at certain points
tm <- c(0.1,0.3,0.5,0.7,0.9)
Est.Lam0t <- PTC.NPMLE.Lam(tm=tm,bet=fit.PTC$res[,1],yobs=yobs,delta=delta,X=X)
print(data.frame(Est=Est.Lam0t,row.names=tm))

}
\references{
PORTIER, F., EL GHOUCH, A. and VAN KEILEGOM, I. (2017). Efficiency and bootstrap in the promotion time cure model. Bernoulli 23 3437–3468.
}
