% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC.R
\name{PTC.NPMLE.InfoM.Half}
\alias{PTC.NPMLE.InfoM.Half}
\title{Half of the information matrix in NPMLE}
\usage{
PTC.NPMLE.InfoM.Half(bet, yobs, delta, X, LinkFunc = list(D0 = exp, D1 = exp))
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}
}
\description{
calculate the half of the information matrix in NPMLE
}
