% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC.R
\name{PTC.NPMLE.Lam}
\alias{PTC.NPMLE.Lam}
\title{Nonparametric component in NPMLE}
\usage{
PTC.NPMLE.Lam(
  tm,
  bet,
  yobs,
  delta,
  X,
  LinkFunc = list(D0 = exp, D1 = exp),
  w = NULL
)
}
\arguments{
\item{tm}{The time points that the nonparametric baseline distribution will be estimated at. The default specification is \code{TRUE}.}

\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{w}{weights for the weighted bootstrap method. The default is \code{NULL}, and it will be treatead as equal weights (\code{w=rep(1,length(yobs))}).}

\item{XI}{a matrix of covariates with 1 as the first column, that is, \code{XI=cbind(1,X)}.}
}
\description{
Nonparametric component for the promotion cure model at specified time points.
}
