% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC.R
\name{PTC.NPMLE.Lambda.Solve}
\alias{PTC.NPMLE.Lambda.Solve}
\title{An auxiliary function for solving lambda given bet}
\usage{
PTC.NPMLE.Lambda.Solve(lam, dRbet, delta, w = NULL)
}
\arguments{
\item{lam}{unknown parameter corresponding to the largrane multiplier.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{w}{weights for the weighted bootstrap method. The default is \code{NULL}, and it will be treatead as equal weights (\code{w=rep(1,length(yobs))}).}

\item{Rbet}{a matrix needed in this function.}
}
\description{
An auxiliary function for solving lambda given bet.
}
