% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC_Online.R
\name{PTC.NPMLE.Renew.ALasso.Batch}
\alias{PTC.NPMLE.Renew.ALasso.Batch}
\title{Online variable selection in promotion time cure model
based on the classical nonparametric maximum likelihood estimator and adaptive Lasso penalty}
\usage{
PTC.NPMLE.Renew.ALasso.Batch(
  yobs,
  delta,
  X,
  initial = TRUE,
  tm = NULL,
  prev = NULL,
  LinkFunc = list(D0 = exp, D1 = exp),
  tunpara.num = 50,
  tunpara.min.ratio = 1e-04,
  tunpara = NULL,
  evatype = c("AIC", "BIC")[2],
  learnrate = c(0.1),
  maxit = 1000,
  eps = 1e-06,
  threshold = 1e-05,
  trace = TRUE,
  Include = TRUE
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{initial}{a logical value. The default specification is \code{TRUE}, indicating that the current data batch is the first batch and no historical summary statistics is available.}

\item{tm}{The time points that the nonparametric baseline distribution will be estimated at. The default specification is \code{TRUE}.}

\item{prev}{indicates the historical summary data. The default specification is \code{NULL} (with \code{initial=TRUE}). Otherwise, it is a list with the following five elements:
\code{bet} is the historical estimated result of regression coefficients;
\code{InfoM} is the historical estimated result of the information matrix;
\code{N} is the sample of historical raw data;
\code{Lamt} is the historical estimated result of the nonparametric baseline distribution at tm (if \code{tm=NULL}, we can set \code{Lamt=NULL} directly);
\code{LamtInfoM} is the historical estimated result of the information matrix for the nonparametric baseline distribution at tm (if \code{tm=NULL}, we can set \code{LamtInfoM=NULL} directly.)}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{tunpara.num}{the number of tuning parameter values that will be fitted at.}

\item{tunpara.min.ratio}{Smallest value for tuning parameter, as a fraction of tunpara.max, the (data derived) entry value (i.e. the smallest value for which all coefficients are zero).}

\item{tunpara}{A user supplied \code{tunpara} sequence. Typical usage is to have the program compute its own \code{tunpara} sequence based on \code{tunpara.num} and \code{tunpara.min.ratio}. Supplying a value of \code{tunpara} overrides this.}

\item{evatype}{the information criterion that will be used in the selection of optimal tuning parameter.}

\item{learnrate}{the learning rate in solving the optimization problem.}

\item{maxit}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default \code{maxit = 5e3}.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{threshold}{a threshold value. The deafult value is \code{1e-5}. If the absolute value of the each component of coefficients is smaller than \code{threshold}, it will be converted into zero.}

\item{trace}{a logical value. The default specification is \code{TRUE}, indicating that several useful information about the current process of fitting will be displayed.}

\item{Include}{a logical value. The default specification is \code{TRUE}, indicating that the recovering term will be included.}
}
\value{
The fitted results are returned (a list). The estimates of regression coefficients is \code{res}.
}
\description{
Conduct online variable selection using promotion time cure model based on the nonparametric maximum likelihood estimation (NPMLE) procedure and adaptive Lasso penalty.
In other words, this function can analyze streaming survival data with a cure fraction and selection the underlying useful variables.
}
\examples{
### ==== An example for doing online variable selection via the promotion time cure model ==== ###

## ---- generate the dataset (full) ---- ##
N <- 4000
pX <- 10
sdata.full  <- sdata.PTC.High(N=N,pX=pX)
yobs  <- sdata.full$sdata$yobs
delta <- sdata.full$sdata$delta
X     <- as.matrix(sdata.full$sdata$X)
rm(sdata.full)

## ---- fit the promotion time cure model in an online manner ---- ##
# prepare basic elements
tm <- c(0.1,0.3,0.5,0.7,0.9)
B <- 20
Nb <- N/B
batch <- ceiling((1:N) / Nb)
Res.Online_ALasso <- list(
  res = array(0,dim=c(B,pX+1,2),
              dimnames=list(paste("batch",1:B,sep=""),paste("bet",0:pX,sep=""),c("EST","SE"))),
  res.Lamt = array(0,dim=c(B,length(tm)),
                   dimnames=list(paste("batch",1:B,sep=""),paste("tm",1:length(tm),sep="")))
)
# the online procedure
for(b in 1:B){  # b <- 1
  cat("Batch",b,"...\n")

  # preparation: the batch idx / the previous elements
  idxb <- batch==b
  if(b==1){ prevb <- NULL }else{
    prevb <- list(
      bet         = fitb$res[,1],
      Score       = fitb$Score,
      InfoM       = fitb$InfoM,
      ScoreAppro  = fitb$ScoreAppro,
      bet.nopen   = fitb$bet.nopen,
      InfoM.nopen = fitb$InfoM.nopen,
      Lamt        = fitb$Lamt,
      LamtInfoM   = fitb$LamtInfoM,
      N           = fitb$N
    )
  }

  # fit the current data batch (with current data the historical statistics)
  fitb <- PTC.NPMLE.Renew.ALasso.Batch(
    yobs=yobs[idxb],delta=delta[idxb],X=as.matrix(X[idxb,,drop=FALSE]),
    initial=(b==1),tm=tm,prev=prevb,Include=TRUE)
  Res.Online_ALasso$res[b,,1] <- fitb$res[,1]
  Res.Online_ALasso$res[b,,2] <- fitb$res[,2]
  Res.Online_ALasso$res.Lamt[b,] <- fitb$res.Lamt[,1]

}
# present the fitted results
print(Res.Online_ALasso$res)
print(Res.Online_ALasso$res.Lamt)

}
