% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC_Online.R
\name{PTC.NPMLE.Renew.Batch}
\alias{PTC.NPMLE.Renew.Batch}
\title{Online estimation in promotion time cure model
based on the classical nonparametric maximum likelihood estimator}
\usage{
PTC.NPMLE.Renew.Batch(
  yobs,
  delta,
  X,
  initial = TRUE,
  tm = NULL,
  prev = NULL,
  LinkFunc = list(D0 = exp, D1 = exp),
  maxit = 1000,
  eps = 1e-06,
  simplify = TRUE
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{initial}{a logical value. The default specification is \code{TRUE}, indicating that the current data batch is the first batch and no historical summary statistics is available.}

\item{tm}{The time points that the nonparametric baseline distribution will be estimated at. The default specification is \code{TRUE}.}

\item{prev}{indicates the historical summary data. The default specification is \code{NULL} (with \code{initial=TRUE}). Otherwise, it is a list with the following five elements:
\code{bet} is the historical estimated result of regression coefficients;
\code{InfoM} is the historical estimated result of the information matrix;
\code{N} is the sample of historical raw data;
\code{Lamt} is the historical estimated result of the nonparametric baseline distribution at tm (if \code{tm=NULL}, we can set \code{Lamt=NULL} directly);
\code{LamtInfoM} is the historical estimated result of the information matrix for the nonparametric baseline distribution at tm (if \code{tm=NULL}, we can set \code{LamtInfoM=NULL} directly.)}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{maxit}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default \code{maxit = 5e3}.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{simplify}{a logical value. The default specification is \code{TRUE}, indicating that the inversion at each Newton iteration step will be avoided (using the inversion of the information matrix from historical data directly).}
}
\value{
The fitted results are returned (a list). The estimates of regression coefficients is \code{res}.
}
\description{
Fit online promotion time cure model based on the nonparametric maximum likelihood estimation (NPMLE) procedure.
In other words, this function can analyze streaming survival data with a cure fraction.
More details about the classical NPMLE method can be found in Portier et al. (2017).
}
\examples{
### ==== An example for fitting the online promotion time cure model ==== ###

## ---- generate the dataset (full) ---- ##
N <- 10000
sdata.full  <- sdata.PTC(N=N)
yobs  <- sdata.full$sdata$yobs
delta <- sdata.full$sdata$delta
X     <- as.matrix(sdata.full$sdata$X)
rm(sdata.full)

## ---- fit the promotion time cure model in an online manner ---- ##
# prepare basic elements
tm <- c(0.1,0.3,0.5,0.7,0.9)
B <- 40
Nb <- N/B
batch <- ceiling((1:N) / Nb)
Res.Online <- list(
  res = array(0,dim=c(B,3,2),
              dimnames=list(paste("batch",1:B,sep=""),paste("bet",0:2,sep=""),c("EST","SE"))),
  res.Lamt = array(0,dim=c(B,length(tm)),
                   dimnames=list(paste("batch",1:B,sep=""),paste("tm",1:length(tm),sep="")))
)
# the online procedure
for(b in 1:B){  # b <- 1
  cat("Batch",b,"...\n")

  # preparation: the batch idx / the previous elements
  idxb <- batch==b
  if(b==1){ prevb <- NULL }else{
    prevb <- list(
      bet       = fitb$res[,1],
      InfoM     = fitb$InfoM,
      Lamt      = fitb$Lamt,
      LamtInfoM = fitb$LamtInfoM,
      N         = fitb$N
    )
  }

  # fit the current data batch (with current data the historical statistics)
  fitb <- PTC.NPMLE.Renew.Batch(
    yobs=yobs[idxb],delta=delta[idxb],X=as.matrix(X[idxb,,drop=FALSE]),
    initial=(b==1),tm=tm,prev=prevb)
  Res.Online$res[b,,1] <- fitb$res[,1]
  Res.Online$res[b,,2] <- fitb$res[,2]
  Res.Online$res.Lamt[b,] <- fitb$res.Lamt[,1]

}
# present the fitted results
print(Res.Online$res)
print(Res.Online$res.Lamt)

}
\references{
PORTIER, F., EL GHOUCH, A. and VAN KEILEGOM, I. (2017). Efficiency and bootstrap in the promotion time cure model. Bernoulli 23 3437–3468.
}
