% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC_Online.R
\name{PTC.NPMLE.Renew.Beta.LogLik}
\alias{PTC.NPMLE.Renew.Beta.LogLik}
\title{Loss function for bet in NPMLE (Online version)}
\usage{
PTC.NPMLE.Renew.Beta.LogLik(
  bet,
  yobs,
  delta,
  XI,
  bet.prev,
  InfoM.prev,
  ScoreAppro.prev,
  N.prev,
  LinkFunc = list(D0 = exp, D1 = exp)
)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{XI}{a matrix of covariates with 1 as the first column, that is, \code{XI=cbind(1,X)}.}

\item{bet.prev}{the historical summary data of regression coefficients.}

\item{InfoM.prev}{the historical summary data of the information matrix (inverse variance-covariance matrix).}

\item{ScoreAppro.prev}{the historical summary data of a term derived from the online variable selection procedure (the recovering term).}

\item{N.prev}{the sample of historical raw data}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}
}
\description{
Online loss function for the promotion cure model (Profiled online log-likelihood function in NPMLE).
Note that it highly depends on the function \code{\link{PTC.NPMLE.Beta.LogLik}} in offline scenario.
}
