% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC.R
\name{PTC.NPMLE.Score.Individual}
\alias{PTC.NPMLE.Score.Individual}
\title{Individual level score vector in NPMLE}
\usage{
PTC.NPMLE.Score.Individual(
  bet,
  yobs,
  delta,
  X,
  LinkFunc = list(D0 = exp, D1 = exp),
  Iscov = TRUE
)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{Iscov}{whether convert the individual level form into a covariance form.}
}
\description{
calculate the individual level score vector in NPMLE
}
