% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC.R
\name{PTC.NPMLE.ScoreInfoM}
\alias{PTC.NPMLE.ScoreInfoM}
\title{Score vector and information matrix in NPMLE}
\usage{
PTC.NPMLE.ScoreInfoM(
  bet,
  yobs,
  delta,
  X,
  LinkFunc = list(D0 = exp, D1 = exp),
  IsScore = FALSE,
  IsInfoM = TRUE
)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{IsScore}{whether the score vector is calculated or not.}

\item{IsInfoM}{whether the information matrix is calculated or not.}
}
\description{
calculate score vector and information matrix in NPMLE.
}
