% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC.R
\name{PTC.NPMLE.Stx}
\alias{PTC.NPMLE.Stx}
\title{Survival function for the whole population (for promotion time cure model)}
\usage{
PTC.NPMLE.Stx(
  yobs,
  delta,
  X,
  bet,
  cross = FALSE,
  tm = NULL,
  LinkFunc = list(D0 = exp, D1 = exp)
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates corresponding to latency part.}

\item{bet}{unknown parameters corresponding to latency part,   or the fitted values of regression coefficients in the latency part returned by the function \code{\link{SMC.PH.fit()}}.}

\item{cross}{a logical value. The default is \code{FALSE}.}

\item{tm}{the time points that the survival function will be calculated at. The default is \code{NULL}.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}
}
\description{
Calculate the survival function for the whole population (for promotion time cure model) at various time points.
}
