% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC_SimData.R
\name{sdata.PTC.High}
\alias{sdata.PTC.High}
\title{The data generating function under the promotion time cure model (high-dimensional)}
\usage{
sdata.PTC.High(
  N,
  bet.noInt = c(0.6, 0.5, 0.4),
  pX = 10,
  X.type = "continuous",
  avalue = -0.5,
  cvalue = 2.2,
  cureRate = NULL,
  censoringRate = NULL,
  corstr = c("IND", "EX", "AR1")[3],
  rho = 0.4,
  tau = 2,
  Lam0t = function(t) {
     ifelse(t > 1, 1, ifelse(t < 0, 0, t))
 },
  LinkFunc = list(D0 = exp, D1 = exp)
)
}
\arguments{
\item{N}{the sample size.}

\item{bet.noInt}{true value of coefficients (not include the intercept). Certain default value is set.}

\item{pX}{the total number of generated covariates. The default value is 10.}

\item{X.type}{the type of covariates. The default type is "continuous" (normal).}

\item{avalue}{a value used to control the cure rate. It is also the intercept term.}

\item{cvalue}{a value used to control the censoring rate.}

\item{cureRate}{desired cure rate.}

\item{censoringRate}{desired censoring rate.}

\item{corstr}{type of correlation structure.}

\item{rho}{the degree of correlation.}

\item{tau}{the truncated point.}

\item{Lam0t}{true baseline distribution function. Certain default value is set.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}
}
\description{
Generate a simulated dataset from the promotion time cure model with relatively high dimensional covariates.
}
