% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC_SimData.R
\name{sdata.PTC}
\alias{sdata.PTC}
\title{The data generating function under the promotion time cure model (low-dimensional)}
\usage{
sdata.PTC(
  N,
  bet = c(-1, 1, 1),
  Lam0t = function(t) {
     ifelse(t > 1, 1, ifelse(t < 0, 0, t))
 },
  LinkFunc = list(D0 = exp, D1 = exp),
  avalue = 0,
  cvalue = 2.5,
  cureRate = NULL,
  censoringRate = NULL,
  Interaction = FALSE
)
}
\arguments{
\item{N}{the sample size.}

\item{bet}{true value of coefficients (include the intercept). Certain default value is set.}

\item{Lam0t}{true baseline distribution function. Certain default value is set.}

\item{LinkFunc}{the link function specified in the promotion time cure model. The default is the exponential function (The proportional hazards cure model).}

\item{avalue}{a value used to control the cure rate.}

\item{cvalue}{a value used to control the censoring rate.}

\item{cureRate}{desired cure rate.}

\item{censoringRate}{desired censoring rate.}

\item{Interaction}{whether to include the interaction term (then bet should be of 4-dimension).}
}
\description{
Generate a simulated dataset from the promotion time cure model with low dimensional covariates.
}
