% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTC_HD.R
\name{Soft.Threshold}
\alias{Soft.Threshold}
\title{The soft-thresholding function}
\usage{
Soft.Threshold(theta, lam)
}
\arguments{
\item{theta}{the parameter that we will calculate at.}

\item{lam}{the inputted tuning parameter}
}
\description{
The soft-thresholding function.
}
